Evaluation result classes shouldn't be nested under the Evaluator classes #1280
1. Moved nested classes Result and CalibratedResult.
2. Refactored to EvaluatorMetrics and BinaryClassificationMetrics
